refactor: split Settings page into modular component files for better readability and maintenance
